pyproject.toml
# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["newsflow/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "UP",
    "YTT",
    # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
    "ASYNC",
    "S",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "DJ",
    "EM",
    "EXE",
    "FA",
    'ISC',
    "ICN",
    "G",
    'INP',
    'PIE',
    "T20",
    'PYI',
    'PT',
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TC",
    "INT",
    # "ARG", # Unused function argument
    "PTH",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    # "NPY",
    # "AIR",
    "PERF",
    # "FURB",
    # "LOG",
    "RUF",
]
ignore = [
    "S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "SIM102", # sometimes it's better to nest
    "UP038", # Checks for uses of isinstance/issubclass that take a tuple
    # of types for comparison.
    # Deactivated because it can make the code slow:
    # https://github.com/astral-sh/ruff/issues/7871
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "UP038",
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
    "coverage==7.10.6",
    "django-coverage-plugin==3.1.1",
    "django-debug-toolbar==6.0.0",
    "django-extensions==4.1",
    "django-stubs[compatible-mypy]==5.2.2",
    "djlint==1.36.4",
    "factory-boy==3.3.2",
    "ipdb==0.13.13",
    "mypy==1.17.1",
    "pre-commit==4.3.0",
    "psycopg[binary]==3.2.9",
    "pytest==8.4.2",
    "pytest-django==4.11.1",
    "pytest-sugar==1.1.1",
    "ruff==0.12.12",
    "sphinx==8.2.3",
    "sphinx-autobuild==2025.8.25",
    "watchfiles==1.1.0",
    "werkzeug[watchdog]==3.1.3",
]

[project]
name = "newsflow"
version = "0.1.0"
description = "is an AI-powered personalized news aggregation platform built with Django that intelligently curates content from 100+ global news sources"
readme = "README.md"

license = { text = "MIT" }
authors = [
  { name = "Abdul Basit Mohammed", email = "abdulbasitm810@gmail.com" },
]
requires-python = "==3.12.*"
dependencies = [
    "argon2-cffi==25.1.0",
    "beautifulsoup4==4.12.3",
    "celery==5.5.3",
    "crispy-tailwind==1.0.3",
    "django==5.1.12",
    "django-allauth[mfa]==65.11.1",
    "pyjwt[crypto]==2.8.0",
    "django-anymail[mailgun]==13.1",
    "django-celery-beat==2.8.1",
    "django-compressor==4.5.1",
    "django-crispy-forms==2.4",
    "django-environ==0.12.0",
    "django-htmx==1.19.0",
    "django-model-utils==5.0.0",
    "django-redis==6.0.0",
    "feedparser==6.0.11",
    "gunicorn==23.0.0",
    "hiredis==3.2.1",
    "lxml==5.3.0",
    "lxml_html_clean==0.4.1",
    "newspaper4k==0.9.3",
    "nltk==3.9.1",
    "pillow==11.3.0",
    "psycopg[c]==3.2.9",
    "python-dateutil==2.9.0",
    "python-json-logger==3.2.1",
    "python-slugify==8.0.4",
    "rcssmin==1.1.2",
    "redis==6.4.0",
    "requests==2.32.3",
    "tldextract==5.1.3",
    "whitenoise==6.9.0",
    "scikit-learn>=1.7.2",
    "pandas>=2.3.2",
    "numpy>=2.3.3",
    "djangorestframework>=3.16.1",
    "transformers>=4.56.2",
    "torch>=2.2.0,<2.3.0",
    "safetensors>=0.4.0",
    "sacremoses>=0.1.1",
    "vadersentiment>=3.3.2",
]
